Move data and profile loads to utils file.
